+++ zendriver/core/util.py
@@ -2,8 +2,10 @@ from __future__ import annotations
 
 import asyncio
 import logging
+import subprocess
 import types
 import typing
+from pathlib import Path
 from typing import Any, Callable, List, Optional, Set, Union
 
 from deprecated import deprecated
@@ -328,19 +330,41 @@ def cdp_get_module(domain: Union[str, types.ModuleType]):
     return domain_mod
 
 
-async def _read_process_stderr(
-    process: asyncio.subprocess.Process, n: int = 2**16
-) -> str:
+def _start_process(
+    exe: str | Path, params: List[str], is_posix: bool
+) -> subprocess.Popen:
+    """
+    Start a subprocess with the given executable and parameters.
+
+    :param exe: The executable to run.
+    :param params: List of parameters to pass to the executable.
+    :param is_posix: Boolean indicating if the system is POSIX compliant.
+
+    :return: An instance of `subprocess.Popen`.
+    """
+    return subprocess.Popen(
+        [str(exe)] + params,
+        stdin=subprocess.PIPE,
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+        close_fds=is_posix,
+    )
+
+
+async def _read_process_stderr(process: subprocess.Popen, n: int = 2**16) -> str:
     """
     Read the given number of bytes from the stderr of the given process.
 
     Read bytes are automatically decoded to utf-8.
     """
-    if process.stderr is None:
-        raise ValueError("Process has no stderr")
+
+    async def read_stderr() -> bytes:
+        if process.stderr is None:
+            raise ValueError("Process has no stderr")
+        return await asyncio.to_thread(process.stderr.read, n)
 
     try:
-        return (await asyncio.wait_for(process.stderr.read(n), 0.25)).decode("utf-8")
+        return (await asyncio.wait_for(read_stderr(), 0.25)).decode("utf-8")
     except asyncio.TimeoutError:
         logger.debug("Timeout reading process stderr")
         return ""